+++ packages/iov-bns/types/context.d.ts
@@ -10,7 +10,7 @@ export interface ChainData {
     readonly chainId: ChainId;
     readonly tickers: Map<string, BcpTicker>;
 }
-/** Like BcpAccount but with no pubkey. Keep compatible to BcpAccount! */
+/** Like BCP's Account but with no pubkey. Keep compatible to Account! */
 export interface WalletData {
     readonly address: Address;
     readonly name?: string;