+++ packages/iov-lisk/src/liskconnection.ts
@@ -4,12 +4,11 @@ import { ReadonlyDate } from "readonly-date";
 import { Producer, Stream } from "xstream";
 
 import {
+  Account,
+  AccountQuery,
+  AddressQuery,
   Algorithm,
-  BcpAccount,
-  BcpAccountQuery,
-  BcpAddressQuery,
   BcpConnection,
-  BcpPubkeyQuery,
   BcpTicker,
   BcpTxQuery,
   BlockHeader,
@@ -22,6 +21,7 @@ import {
   Nonce,
   PostableBytes,
   PostTxResponse,
+  PubkeyQuery,
   PublicKeyBundle,
   PublicKeyBytes,
   TokenTicker,
@@ -164,7 +164,7 @@ export class LiskConnection implements BcpConnection {
     ];
   }
 
-  public async getAccount(query: BcpAccountQuery): Promise<BcpAccount | undefined> {
+  public async getAccount(query: AccountQuery): Promise<Account | undefined> {
     const address = isPubkeyQuery(query) ? pubkeyToAddress(query.pubkey.data) : query.address;
 
     const url = this.baseUrl + `/api/accounts?address=${address}`;
@@ -172,8 +172,8 @@ export class LiskConnection implements BcpConnection {
     const responseBody = result.data;
 
     // here we are expecting 0 or 1 results
-    const accounts: ReadonlyArray<BcpAccount> = responseBody.data.map(
-      (responseItem: any): BcpAccount => {
+    const accounts: ReadonlyArray<Account> = responseBody.data.map(
+      (responseItem: any): Account => {
         const itemBalance: unknown = responseItem.balance;
         const itemPubkey: unknown = responseItem.publicKey;
 
@@ -207,11 +207,11 @@ export class LiskConnection implements BcpConnection {
     return accounts.length > 0 ? accounts[0] : undefined;
   }
 
-  public async getNonce(_: BcpAddressQuery | BcpPubkeyQuery): Promise<Nonce> {
+  public async getNonce(_: AddressQuery | PubkeyQuery): Promise<Nonce> {
     return generateNonce();
   }
 
-  public async getNonces(_: BcpAddressQuery | BcpPubkeyQuery, count: number): Promise<ReadonlyArray<Nonce>> {
+  public async getNonces(_: AddressQuery | PubkeyQuery, count: number): Promise<ReadonlyArray<Nonce>> {
     const checkedCount = new Uint53(count).toNumber();
     // use unique nonces to ensure the same transaction content leads to a different transaction ID
     // [now-3, now-2, now-1, now] for 4 nonces
@@ -221,10 +221,10 @@ export class LiskConnection implements BcpConnection {
     });
   }
 
-  public watchAccount(query: BcpAccountQuery): Stream<BcpAccount | undefined> {
+  public watchAccount(query: AccountQuery): Stream<Account | undefined> {
     let lastEvent: any = {}; // default to a dummy value to ensure an initial undefined event is sent
     let pollInternal: NodeJS.Timeout | undefined;
-    const producer: Producer<BcpAccount | undefined> = {
+    const producer: Producer<Account | undefined> = {
       start: listener => {
         const poll = async () => {
           try {